+++ server/api.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
-import user_router from "./router/userRouter";
-import org_router from "./router/orgRouter";
+import user_router from "@user/router/userRouter";
+import org_router from "@org/router/orgRouter";
 
 const api_router = Router();
