+++ server/package.json
@@ -8,6 +8,7 @@
     "dev": "concurrently \"tsc --watch\" \"nodemon ./dist/index.js\" "
   },
   "dependencies": {
+    "@types/bcrypt": "^5.0.0",
     "bcrypt": "^5.1.0",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
@@ -15,8 +16,8 @@
     "mongoose": "^6.7.2"
   },
   "devDependencies": {
-    "@types/express": "^4.17.14",
     "@types/cors": "^2.8.5",
+    "@types/express": "^4.17.14",
     "@types/node": "^18.11.9",
     "concurrently": "^7.6.0",
     "nodemon": "^2.0.20",
