+++ server/tsconfig.json
@@ -9,9 +9,14 @@
     /* Modules */
     "module": "commonjs" /* Specify what module code is generated. */,
     // "moduleResolution": "node",                       /* Specify how TypeScript looks up a file from a given module specifier. */
-    // "baseUrl": "./",                                  /* Specify the base directory to resolve non-relative module names. */
-    // "paths": {},                                      /* Specify a set of entries that re-map imports to additional lookup locations. */
-    // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
+    "baseUrl": "./" /* Specify the base directory to resolve non-relative module names. */,
+    "paths": {
+      "@user/*": ["./user/*"],
+      "@org/*": ["./org/*"]
+    } /* Specify a set of entries that re-map imports to additional lookup locations. */,
+    // "typeRoots": [
+    //   "../client/src/types/global.d.ts"
+    // ] /* Specify multiple folders that act like './node_modules/@types'. */,
     "resolveJsonModule": true /* Enable importing .json files. */,
 
     /* Emit */
