+++ server/types/global.d.ts
@@ -1,0 +1,12 @@
+export {};
+
+declare global {
+  interface Project {
+    id: string;
+    tasks: any[];
+    name: string;
+    description?: string | undefined;
+    favorite?: boolean | undefined;
+    last_save: number;
+  }
+}
