--- server/types/project.ts
@@ -1,8 +1,0 @@
-export interface Project {
-  id: string;
-  tasks: any[];
-  name: string;
-  description?: string | undefined;
-  favorite?: boolean | undefined;
-  last_save: number;
-}
